+++ Lib.Redist/Symbols/Token.cs
@@ -17,6 +17,6 @@ namespace Hime.Redist.Symbols
         /// </summary>
         /// <param name="sid">ID of the terminal matched by this Token</param>
         /// <param name="name">Name of the terminal matched by this Token</param>
-        protected Token(ushort sid, string name): base(sid, name) { }
+        protected Token(int sid, string name): base(sid, name) { }
     }
 }
